table: fix warnings
Warnings revealed we were passing a Column instance rather than a dtype,
and it only worked before because Column has an attribute called 'dtype'.

DeprecationWarning: in the future the `.dtype` attribute of a given
datatype object must be a valid dtype instance. `data_type.dtype` may
need to be coerced using `np.dtype(data_type.dtype)`. (Deprecated NumPy
1.20)

+++ python/pfs/drp/stella/table.py
@@ -97,7 +97,7 @@ class Table:
         """
         return cls.fromColumns(
             **{
-                name: np.array([getattr(ll, name) for ll in rows], dtype=dtype)
+                name: np.array([getattr(ll, name) for ll in rows], dtype=dtype.dtype)
                 for name, dtype in cls._schema.items()
             }
         )
@@ -113,7 +113,9 @@ class Table:
             and each column must be ``array_like``.
         """
         return cls(
-            DataFrame({name: np.array(columns[name], dtype=dtype) for name, dtype in cls._schema.items()})
+            DataFrame(
+                {name: np.array(columns[name], dtype=dtype.dtype) for name, dtype in cls._schema.items()}
+            )
         )
 
     @classmethod
